+++ .github/workflows/ReleaseNotes.md
@@ -1,6 +1,9 @@
+# Release Notes
+
 ## What's Changed
-* Add command switch to disable automatic overwriting as requested in [#30](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/30)
-* Added paging support to help relieve sever stress as requested in [#6](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/6)
+
+* Add library command line switch to restrict to only specified libraries per [#38](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/38)
 
 ## Known Bugs
+
 * None! If you encounter any bugs, please open an [issue](https://github.com/johnkiddjr/PlexMatch-File-Generator/issues/new)!